Make the QuestionAnswerer support short queries
In case we want to support short queries like "cancer", we have to make
the QuesionAnswerer implementation more tolerant. In particular, it now
can return the subtree even at the depth=1 if it matches the query. The
null value is now returned only if the search fails at the root node.